feat:i 18next and <burger/> being in one component
